+++ pymtheg.py
@@ -1,10 +1,11 @@
 """
 pymtheg: A Python script to share songs from Spotify/YouTube as a 15 second clip
--------
+
+--------------------------------------------------------------------------------
 
 This is free and unencumbered software released into the public domain.
 
------------------------------------------------------------------------------------------
+-----------------------------------------------------------------------
 
 Anyone is free to copy, modify, publish, use, compile, sell, or distribute this software,
 either in source code form or as a compiled binary, for any purpose, commercial or
@@ -25,8 +26,7 @@ FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN 
 For more information, please refer to <http://unlicense.org/>
 """
 
-from re import M
-from typing import Iterable, List, NamedTuple, Optional, Tuple, Union
+from typing import Iterable, List, Literal, NamedTuple, Optional, Tuple, Union
 
 from argparse import ArgumentParser, RawTextHelpFormatter
 from tempfile import TemporaryDirectory
@@ -47,26 +47,34 @@ FFARGS: str = (
 )
 OUT: str = "{artists} - {title}"
 TIMESTAMP_FORMAT: str = " ({cs}{cer})"
+CLIP_START: str = "0"
+CLIP_END: str = "+15"
 
 premsg_info = "[dim]pymtheg: [/dim][bold cyan]info[/bold cyan][dim]:[/]"
 premsg_error = "[dim]pymtheg: [/dim][bold red]error[/bold red][dim]:[/]"
 
 
-class EndTimestamp(NamedTuple):
+class Timestamp(NamedTuple):
     """
-    end timestamp named tuple
+    timestamp named tuple
 
+    type: Literal[0] | Literal[1]
+        0 if start timestamp; 1 if end timestamp
     ss: int
-        timestamp of clip end in seconds
+        timestamp in seconds
+    random: bool
+        is timestamp random
     relative: bool = False
-        is ss relative to clip start?
+        is timestamp relative
     """
 
-    ts: str
+    type: Union[Literal[0], Literal[1]]
+    ss: int
+    random: bool = False
     relative: bool = False
 
     def __str__(self) -> str:
-        return self.ts
+        return ("+" if self.relative else "") + str(self.ss)
 
 
 class Behaviour(NamedTuple):
@@ -79,8 +87,8 @@ class Behaviour(NamedTuple):
     ext: str
     sdargs: List[str]
     ffargs: List[str]
-    clip_start: str
-    clip_end: EndTimestamp
+    clip_start: Timestamp
+    clip_end: Timestamp
     image: Optional[Path]
     use_defaults: bool
     yes: bool
@@ -114,26 +122,44 @@ def main() -> None:
             if song_path.suffix not in [".m4a", ".ogg", ".flac", ".mp3", ".wav", ".opus"]:
                 continue
 
+            # duration retrieval
+            with console.status(f"[dim]status: probe song duration[/]", spinner="arc"):
+                proc = invocate(
+                    console=console,
+                    name="ffprobe",
+                    args=[
+                        "-print_format",
+                        "json",
+                        "-show_entries",
+                        "format=duration",
+                        song_path,
+                    ],
+                    capture_output=True,
+                )
+                song_duration: int = int(
+                    loads(proc.stdout)["format"]["duration"].split(".")[0]
+                )
+
             if processed == 0:
                 # print timestamp format/using default message on first song
                 if bev.use_defaults:
                     console.print(
-                        f"{premsg_info} using defaults, clip start will be {bev.clip_start} and clip "
-                        f" end will be {bev.clip_end}\n"
+                        f'{premsg_info} using defaults, clip start is "{bev.clip_start}"'
+                        f' and clip end is "{bev.clip_end}"\n'
                     )
 
                 else:
                     console.print(f"{premsg_info} enter timestamps in format \[hh:mm:]ss")
-                    console.print("               timestamp can be '*' for random")
+                    console.print('               timestamp can be "*" for random')
                     console.print(
-                        "               end timestamp can be relative, prefix with '+'"
+                        '               end timestamp can be relative, prefix with "+"'
                     )
                     console.print(
                         f"               press enter to use given defaults "
-                        f"({bev.clip_start}, {bev.clip_end})\n"
+                        f'("{bev.clip_start}", "{bev.clip_end}")\n'
                     )
 
-            console.print(f"- [bold]{song_path.stem}[/]")
+            console.print(f"- [bold]{song_path.stem}[/] ({to_timestamp(song_duration)})")
 
             # generate query/info messages
             _msg_format = "    {}: "
@@ -168,26 +194,6 @@ def main() -> None:
             info_notice = _msg_format.format(_info_notice.rjust(_longest_msg_len))
             indent = len(_msg_format) - 2 + _longest_msg_len
 
-            # duration retrieval
-            with console.status(
-                f"[dim]{info_status}probe song duration[/]", spinner="arc"
-            ):
-                proc = invocate(
-                    console=console,
-                    name="ffprobe",
-                    args=[
-                        "-print_format",
-                        "json",
-                        "-show_entries",
-                        "format=duration",
-                        song_path,
-                    ],
-                    capture_output=True,
-                )
-                song_duration: int = int(
-                    loads(proc.stdout)["format"]["duration"].split(".")[0]
-                )
-
             # construct working paths
             song_path = song_path.absolute()
             song_clip_path = tmpdir.joinpath(f"{song_path.stem}_clip.mp3").absolute()
@@ -195,16 +201,9 @@ def main() -> None:
             video_clip_path = tmpdir.joinpath(f"{song_path.stem}_clip.mp4").absolute()
 
             # get timestamps
-            start_timestamp = parse_timestamp(
-                ts=bev.clip_start, song_duration=song_duration
+            start_timestamp, end_timestamp = parse_timestamps(
+                bev.clip_start, bev.clip_end, duration=song_duration
             )
-            assert isinstance(start_timestamp, int)
-            end_timestamp = parse_timestamp(
-                ts=bev.clip_end.ts,
-                song_duration=song_duration,
-                relative_to=start_timestamp,
-            )
-            assert isinstance(end_timestamp, int)
 
             if bev.clip_end.relative:
                 end_timestamp += start_timestamp
@@ -213,15 +212,17 @@ def main() -> None:
                 end_timestamp = song_duration
 
             if not bev.use_defaults:
+                # timestamp prompt
                 while True:
+                    _start_timestamp: Optional[Timestamp] = None
+                    _end_timestamp: Optional[Timestamp] = None
+
                     # starting timestamp
                     while True:
                         cs_response = input(query_clip_start)
 
                         if cs_response != "":
-                            _start_timestamp = parse_timestamp(
-                                cs_response, song_duration=song_duration
-                            )
+                            _start_timestamp = check_timestamp(0, cs_response)
 
                             if _start_timestamp is None:
                                 # invalid format
@@ -232,56 +233,60 @@ def main() -> None:
                                     + "[/dim][bold] invalid timestamp",
                                 )
 
-                            elif _start_timestamp >= song_duration:
-                                # invalid, timestamp >= song duration
-                                console.print(
-                                    "[dim][red]"
-                                    + (" " * indent)
-                                    + ("^" * len(cs_response))
-                                    + "[/dim][bold] timestamp exceeds song duration",
-                                )
-
                             else:
-                                # valid, continue
-                                end_timestamp = (
-                                    end_timestamp - start_timestamp
-                                ) + _start_timestamp
-                                start_timestamp = _start_timestamp
-                                break
+                                if _start_timestamp.ss > song_duration:
+                                    # invalid, timestamp >= song duration
+                                    console.print(
+                                        "[dim][red]"
+                                        + (" " * indent)
+                                        + ("^" * len(cs_response))
+                                        + "[/dim][bold] timestamp exceeds song duration",
+                                    )
 
-                        elif cs_response == "":
+                                else:
+                                    break
+
+                        else:
+                            _start_timestamp = bev.clip_start
                             break
 
                     # ending timestamp
                     while True:
                         ce_response = input(query_clip_end)
+
                         if ce_response != "":
-                            _end_timestamp = parse_timestamp(
-                                ce_response,
-                                relative_to=start_timestamp,
-                                song_duration=song_duration,
-                            )
+                            _end_timestamp = check_timestamp(1, ce_response)
 
                             if _end_timestamp is None:
-                                # reprompt if invalid
+                                # invalid format
                                 console.print(
-                                    (" " * indent) + ("^" * len(ce_response)),
-                                    "invalid timestamp",
+                                    "[dim][red]"
+                                    + (" " * indent)
+                                    + ("^" * len(cs_response))
+                                    + "[/dim][bold] invalid timestamp",
                                 )
 
                             else:
-                                end_timestamp = _end_timestamp
                                 break
 
-                        elif ce_response == "":
+                        else:
+                            _end_timestamp = bev.clip_end
                             break
+
+                    assert isinstance(_start_timestamp, Timestamp)  # type: ignore
+                    assert isinstance(_end_timestamp, Timestamp)  # type: ignore
+
+                    # parse timestamps
+                    start_timestamp, end_timestamp = parse_timestamps(
+                        _start_timestamp, _end_timestamp, duration=song_duration
+                    )
 
                     # confirm timestamps
                     if bev.yes:
                         break
 
                     # dont prompt confirmation if defaults were used
-                    if not cs_response == "" and ce_response == "":
+                    if not (cs_response == "" and ce_response == ""):
                         console.print(
                             "{premsg}clip duration: {start} -> {end} ({duration}s)".format(
                                 premsg=info_notice,
@@ -303,6 +308,11 @@ def main() -> None:
                     else:
                         break
 
+            elif start_timestamp > song_duration:
+                console.print(f"{info_notice}skipping song")
+                processed += 1
+                break
+
             # construct and confirm output path
             out_path: Path = bev.dir.joinpath(
                 "{name}{timestamp}.{ext}".format(
@@ -321,7 +331,7 @@ def main() -> None:
                 out_path.exists()
                 and bev.yes is False
             ):
-                console.print(f"{info_notice}'{out_path.name}' exists in output dir.")
+                console.print(f'{info_notice}"{out_path.name}" exists in output dir.')
                 overwrite_response = input(
                     f"{' ' * indent}overwrite? ([y]es/[n]o/[c]hange) "
                 ).lower()
@@ -428,7 +438,8 @@ def main() -> None:
 
 
 def part_of_day() -> str:
-    """used to greet user goodbye
+    """
+    used to greet user goodbye
 
     call it bloat or whatever, i like it
     """
@@ -444,68 +455,108 @@ def part_of_day() -> str:
     )
 
 
-def parse_timestamp(
-    ts: str,
-    song_duration: int,
-    relative_to: Optional[int] = None,
-) -> Optional[int]:
+def check_timestamp(type: Union[Literal[0], Literal[1]], ts: str) -> Optional[Timestamp]:
     """
-    parse user-submitted timestamp
+    checks timestamps for timestamp retrieval and command line argument validation
 
     ts: str
-        timestamp following [hh:mm:]ss format (e.g. 2:49, 5:18:18)
-    song_duration: int
-        used for the -1 relative end timestamp or random timestamps
-    relative_to: Optional[int] = None
-        used for relative end or random timestamps
+        timestamp string
+    type: Literal[0] | Literal[1]
+        0 if start timestamp; 1 if end timestamp
+
+    returns a Timestamp object if check was successful else None
     """
-    timestamp = 0
+    ts = ts.strip()
 
     if ts == "*":
-        if song_duration == -1:
-            return -1
+        return Timestamp(type=type, ss=0, random=True)
 
-        return randint(0 if relative_to is None else relative_to + 1, song_duration)
-
-    elif ts == "-1" and song_duration is not None:
-        return song_duration
-
-    elif ts.startswith("+") and relative_to is not None:
-        ts = ts[1:]
-        timestamp += relative_to
-
-    sts = ts.split(":")  # split time stamp (hh:mm:ss)
-    sts.reverse()  # (ss:mm:hh)
-
-    tu_conv = [1, 60, 3600]  # time unit conversion
-    total_ss = 0  # total seconds
-
-    if len(sts) < 4:
-        for tu, tu_c in zip(sts, tu_conv):
-            if tu.isnumeric():
-                total_ss += int(tu) * tu_c
-
-            else:
-                return None
-
-        return timestamp + total_ss
+    elif ts == "-1":
+        return Timestamp(type=type, ss=-1)
 
     else:
-        return None
+        relative: bool
+        if ts.startswith("+"):
+            if type == 0:  # relative timestamps in start timestamp are not allowed
+                return None
+
+            relative = True
+            ts = ts[1:]
+
+        else:
+            relative = False
+
+        sts = ts.split(":")  # split time stamp (hh:mm:ss)
+        sts.reverse()  # (ss:mm:hh)
+
+        tu_conv = [1, 60, 3600]  # time unit conversion
+        total_ss = 0  # total seconds
+
+        if len(sts) < 4:
+            for tu, tu_c in zip(sts, tu_conv):
+                if tu.isnumeric():
+                    total_ss += int(tu) * tu_c
+
+                else:
+                    return None
+
+            return Timestamp(type=type, ss=total_ss, relative=relative)
+
+        else:
+            return None
 
 
-def to_timestamp(timestamp: int) -> str:
-    """returns a [hh:mm:]ss timestamp string from `timestamp: int`"""
-    _mm = timestamp // 60
+def parse_timestamps(start: Timestamp, end: Timestamp, duration: int) -> Tuple[int, int]:
+    """
+    parses start timestamp and end timestamp into absolute seconds
+
+    start: Timestamp
+        start timestamp
+    end: Timestamp
+        end timestamp
+    duration: int
+        song duration in seconds
+
+    returns song start and song end respectively, in seconds
+    """
+    ts_start: int
+    ts_end: int
+
+    if start.random and end.random:
+        ts_start = randint(0, duration)
+        ts_end = randint(ts_start, duration)
+
+    elif start.random:
+        ensure_random = end.ss if end.relative else 0
+        ts_start = randint(0, duration - ensure_random)
+        ts_end = ts_start + end.ss
+
+    elif end.random:
+        ts_start = start.ss
+        ts_end = randint(start.ss, duration)
+
+    else:
+        ts_start = start.ss
+        ts_end = (start.ss + end.ss) if end.relative else end.ss
+
+    return (ts_start, ts_end)
+
+
+def to_timestamp(ts: int) -> str:
+    """returns a [(h*):mm:]ss timestamp string from `ts: int`"""
+    _mm = ts // 60
     hh = _mm // 60
     mm = _mm - hh * 60
-    ss = timestamp % 60
+    ss = ts % 60
 
-    return ":".join([str(unit) for unit in (hh, mm) if unit != 0] + [str(ss)])
+    return ":".join(
+        [str(u).rjust(2, "0") for u in (hh, mm) if u != 0] + [str(ss).rjust(2, "0")]
+    ).lstrip("0")
 
 
 def tf_format(string: str, clip_start: int, clip_end: int) -> str:
-    """formats a string with clip information, returns result
+    """
+    formats a string with clip information, returns result
 
     clip_start: int
         clip start in seconds
@@ -528,7 +579,7 @@ def tf_format(string: str, clip_start: int, clip_end: int) -> str:
             else:
                 result = str(unit) + result
 
-        return result
+        return result.lstrip("0")
 
     replaceables = (
         ("{cs}", ts_format(clip_start)),
@@ -624,7 +675,7 @@ def get_args(console: Console) -> Behaviour:
         description=(
             "a python script to share songs from Spotify/YouTube as a 15 second clip"
         ),
-        epilog=f'''querying:
+        epilog=f"""querying:
   queries are passed onto spotdl, and thus must be any one of the following:
     1. text
       "<query>"
@@ -642,7 +693,7 @@ argument defaults:
   -o, --out:
     "{OUT}"
   -t, --timestamp-format:
-    {TIMESTAMP_FORMAT}
+    "{TIMESTAMP_FORMAT}"
 
 formatting:
   available placeholders:
@@ -650,21 +701,40 @@ formatting:
       {{artist}}, {{artists}}, {{title}}, {{album}}, {{playlist}}
     from pymtheg:
       {{cs}}
-        clip end as per [(h*)mm]ss, e.g. 10648 (1h, 06m, 48s)
+        clip end as per [(h*)mm]ss
+        e.g. 10648 (1h, 06m, 48s)
       {{css}}
-        clip end in seconds, e.g. 4008 (1h, 6m, 48s -> 4008s)
+        clip end in seconds
+        e.g. 4008 (1h, 6m, 48s -> 4008s)
       {{ce}}
         clip end as per [(h*)mm]ss, e.g. 10703 (1h, 07m, 03s)
       {{ces}}
-        clip end in seconds, e.g. 4023 (1h, 07m, 03s -> 4023s)
+        clip end in seconds
+        e.g. 4023 (1h, 07m, 03s -> 4023s)
       {{cer}}
+        clip end relative to clip start, prefixed with +
         e.g. +15
     
       notes:
         1. pymtheg placeholders can only be used with `-tf, --timestamp-format`
         2. "[(h*)mm]ss": seconds and minutes will always be represented as 2
-           digits and will be right adjusted with 0s if needed, however hours
-           can be represented by any number of characters, e.g. "1" or "123456"''',
+           digits and will be right adjusted with 0s if needed, unless they are
+           the first shown unit where they may have up to two characters. hours
+           can be represented by any number of characters.
+           e.g. "138:02:09", "1:59:08", "2:05", "6"
+
+examples:
+  1. get a song through a spotify link
+    pymtheg "https://open.spotify.com/track/..."
+  2. get a song through a search query
+    pymtheg "thundercat - them changes"
+  3. get multiple songs through multiple queries
+    pymtheg "https://open.spotify.com/track/..." "<query 2>"
+  4. get a random 15s clip of a song
+    pymtheg "<query>" -cs "*" -ce "+15" -ud 
+
+  note: see querying for more information on queries
+""",
         formatter_class=RawTextHelpFormatter,
     )
 
@@ -710,7 +780,7 @@ formatting:
         help="specify clip start (default 0)",
         dest="clip_start",
         type=str,
-        default="0",
+        default=CLIP_START,
     )
     parser.add_argument(
         "-ce",
@@ -718,7 +788,7 @@ formatting:
         help="specify clip end (default +15)",
         dest="clip_end",
         type=str,
-        default="+15",
+        default=CLIP_END,
     )
     parser.add_argument(
         "-i", "--image", help="specify custom image", type=Path, default=None
@@ -726,7 +796,7 @@ formatting:
     parser.add_argument(
         "-ud",
         "--use-defaults",
-        help="use 0 as clip start and --clip-length as clip end",
+        help="use --clip-start as clip start and --clip-length as clip end",
         dest="use_defaults",
         action="store_true",
         default=False,
@@ -742,20 +812,17 @@ formatting:
     args = parser.parse_args()
 
     # validate clip start/end
-    if args.clip_start != "*" and parse_timestamp(args.clip_start, -1) is None:
-        console.print(f"{premsg_error} invalid clip start (format: [hh:mm:]ss)")
+    start_timestamp = check_timestamp(0, args.clip_start)
+    end_timestamp = check_timestamp(1, args.clip_end)
+
+    if start_timestamp is None:
+        console.print(f"{premsg_error} invalid clip start (format: \[hh:mm:]ss)")
         exit(1)
 
-    end_timestamp = parse_timestamp(
-        # use dumb values, because who knows if somebody sets it to -1
-        args.clip_end,
-        song_duration=-1,
-        relative_to=0,
-    )
     if end_timestamp is None:
         console.print(
-            f"{premsg_error} invalid clip end (format: [hh:mm:]ss), "
-            "prefix with '+' for relative timestamp"
+            f"{premsg_error} invalid clip end (format: \[hh:mm:]ss), "
+            'prefix with "+" for relative timestamp'
         )
         exit(1)
 
@@ -799,10 +866,8 @@ formatting:
         ext=args.ext,
         sdargs=args.sdargs.split(),
         ffargs=args.ffargs.split(),
-        clip_start=args.clip_start,
-        clip_end=EndTimestamp(
-            args.clip_end, relative=True if args.clip_end.startswith("+") else False
-        ),
+        clip_start=start_timestamp,
+        clip_end=end_timestamp,
         image=args.image,
         use_defaults=args.use_defaults,
         yes=args.yes,